+++ common/src/main/java/dev/terminalmc/chatnotify/command/Commands.java
@@ -23,10 +23,14 @@ import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import dev.terminalmc.chatnotify.ChatNotify;
 import dev.terminalmc.chatnotify.gui.screen.RootScreen;
 import dev.terminalmc.chatnotify.util.Unicode;
+import dev.terminalmc.chatnotify.util.text.DebugParseUtil;
 import net.minecraft.client.Minecraft;
 import net.minecraft.commands.CommandBuildContext;
 import net.minecraft.network.chat.Component;
 
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.List;
 
 import static net.minecraft.commands.Commands.argument;
@@ -71,27 +75,74 @@ public class Commands {
                     mc.schedule(() -> mc.setScreen(new RootScreen(mc.screen)));
                     return Command.SINGLE_SUCCESS;
                 })
-                .then(literal("format")
-                        .then(argument("pattern", StringArgumentType.greedyString())
-                                .suggests(((ctx, builder) -> {
-                                    String[] split = ctx.getInput().split("format ");
-                                    String input = split.length < 2 ? "" : split[1];
-                                    String pre = input.endsWith("$") ? input : input + "$";
-                                    FORMAT_CODES.forEach((s) -> builder.suggest(pre + s));
-                                    return builder.buildFuture();
-                                }))
-                                .executes(ctx -> {
-                                    String pattern = StringArgumentType.getString(ctx, "pattern");
-                                    Component text = Component.literal(pattern.replaceAll(
-                                            "\\$",
-                                            Unicode.SECTION.str
-                                    ));
+                .then(literal("debug")
+                        .then(literal("format")
+                                        .then(argument("pattern", StringArgumentType.greedyString())
+                                                .suggests(((ctx, builder) -> {
+                                                    String[] split = ctx.getInput().split("format ");
+                                                    String input = split.length < 2 ? "" : split[1];
+                                                    String pre = input.endsWith("$") ? input : input + "$";
+                                                    FORMAT_CODES.forEach((s) -> builder.suggest(pre + s));
+                                                    return builder.buildFuture();
+                                                }))
+                                                .executes(ctx -> {
+                                                    String pattern = StringArgumentType.getString(
+                                                            ctx, "pattern"
+                                                    );
+                                                    Component text = Component.literal(pattern.replaceAll(
+                                                            "\\$",
+                                                            Unicode.SECTION.str
+                                                    ));
 
                                     mc.gui.getChat().addClientSystemMessage(text);
                                     return Command.SINGLE_SUCCESS;
                                 })
                         )
+                        )
+                        .then(literal("parse")
+                                .then(literal("string")
+                                        .then(argument("string", StringArgumentType.greedyString())
+                                                .executes(ctx -> {
+                                                    String string =
+                                                            StringArgumentType.getString(
+                                                                    ctx,
+                                                                    "string"
+                                                            );
+                                                    Component text =
+                                                            DebugParseUtil.parseMutableComponent(string);
+
+                                                    mc.gui.getChat().addClientSystemMessage(text);
+                                                    return Command.SINGLE_SUCCESS;
+                                                })
+                                        )
+                                )
+                                .then(literal("file")
+                                        .then(argument("path", StringArgumentType.greedyString())
+                                                .executes(ctx -> {
+                                                    String path = StringArgumentType.getString(
+                                                            ctx,
+                                                            "path"
+                                                    );
+                                                    parseFromPath(mc, path);
+
+                                                    return Command.SINGLE_SUCCESS;
+                                                })
+                                        )
+                                )
+                        )
                 )
         );
+    }
+
+    private static void parseFromPath(Minecraft mc, String path) {
+        try {
+            List<String> lines = Files.readAllLines(Path.of(path));
+            for (String line : lines) {
+                Component text = DebugParseUtil.parseMutableComponent(line.strip());
+                mc.gui.getChat().addClientSystemMessage(text);
+            }
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
     }
 }